+++ aws-sdk-core/lib/aws-sdk-core/client.rb
@@ -9,6 +9,7 @@ module Aws
       'Aws::Plugins::ParamConverter',
       'Aws::Plugins::ParamValidator',
       'Aws::Plugins::UserAgent',
+      'Aws::Plugins::HelpfulSocketErrors',
       'Aws::Plugins::RetryErrors',
       'Aws::Plugins::GlobalConfiguration',
       'Aws::Plugins::RegionalEndpoint',